Fixest 0.13 compatibilty (#1036)
* bump to fixest 0.13 for testing

* Update inference defaults to iid close #1030

* change collin tol defaults to 1e-09

* stringmagic 1.2.0 requirement

* rebuild lock

* refactor

* refactor

* move default env installation to conda to make sure packages can be installed together

* update deps

* follow fixef_rm defaults

* update fixef_rm arg

* fixest renames dof.K to df.K

* initial move to pyproject toml

* pyproject

* all envs work

* finish migration to pixi.toml?

* jax for dev

* rename fixef_k to k_fixef

* rename cluster_df

* rename all ssc args

* bring back pytest markers

* inf_na_droping fix

* fix prediction tests

* fix new fixef level bug

* add test exception as fixest solved bug with 0.13

* fix wildbootest error introduced

* adjust rwolf code due to new inference defaults

* also update vcov defaults

* also update vcov defaults

* two way clustering test refactor

* create pytest ini

* backwards compatibility

* update changelog

* fix

* wildboottest changelog

* delete config file

* no drop_infinite coef

* fix mypy

* Rename back to separation check

* More separation check fixes

* Update FixestMulti_.py

* Separation check

* separation

* separation

* Separation

* separation

* Singleton

* Singleton

* Separation

* separation_check

* delete scripts folder in tests

* delete conftest

* undo

* delete file

* cleanups

* proper use pixi toml

* bring back pyproject toml

* update pixi toml

* update lock file

* fix?

* appease mypy

* appease mypy

* fix stringmagic

* stringmagic dep

* R install back

* fix changelog bug

* more updates

* more updates

* update quartodoc

* tweaks

* update contributing md

* change ssc defaults for heteroskedastic errors

* finally pass all tests

+++ pyfixest/did/lpdid.py
@@ -339,7 +339,7 @@ def _pooled_adjustment(
 
     # Calculate lead sum
     for k in range(0, pool_lead + 1, 1):
-        x += df.groupby(idname)[y].shift(-k).to_numpy()
+        x += df.groupby(idname)[y].shift(-k).to_numpy().astype(float)
 
     # Average the lead sum
     x /= pool_lead + 1
